requirements.txt
openai>=0.27.0
PyPDF2>=3.0.0
numpy>=1.20.0
python-dotenv>=0.21.0
tiktoken>=0.3.0
faiss-cpu>=1.7.1